requirements.txt
fastapi==0.135.1
python-dotenv==1.2.2
google-generativeai==0.8.4
sentence-transformers==5.2.3
joblib==1.5.3
pandas==3.0.1
scikit-learn==1.8.0
uvicorn==0.42.0
python-multipart==0.0.22
